refactor(core): Refactor webhooks
