+++ packages/cli/src/AbstractServer.ts
@@ -13,11 +13,9 @@ import * as Db from '@/Db';
 import { N8nInstanceType } from '@/Interfaces';
 import type { IExternalHooksClass } from '@/Interfaces';
 import { ExternalHooks } from '@/ExternalHooks';
-import { send, sendErrorResponse, ServiceUnavailableError } from '@/ResponseHelper';
+import { sendErrorResponse, ServiceUnavailableError } from '@/ResponseHelper';
 import { rawBodyReader, bodyParser, corsMiddleware } from '@/middlewares';
-import { TestWebhooks } from '@/TestWebhooks';
-import { WaitingWebhooks } from '@/WaitingWebhooks';
-import { webhookRequestHandler } from '@/WebhookHelpers';
+import { ActiveWebhooks, TestWebhooks, WaitingWebhooks } from '@/webhooks';
 import { generateHostInstanceId } from './databases/utils/generators';
 
 export abstract class AbstractServer {
@@ -39,12 +37,6 @@ export abstract class AbstractServer {
 
 	protected restEndpoint: string;
 
-	protected endpointWebhook: string;
-
-	protected endpointWebhookTest: string;
-
-	protected endpointWebhookWaiting: string;
-
 	protected instanceId = '';
 
 	protected webhooksEnabled = true;
@@ -62,11 +54,7 @@ export abstract class AbstractServer {
 		this.sslCert = config.getEnv('ssl_cert');
 
 		this.timezone = config.getEnv('generic.timezone');
-
 		this.restEndpoint = config.getEnv('endpoints.rest');
-		this.endpointWebhook = config.getEnv('endpoints.webhook');
-		this.endpointWebhookTest = config.getEnv('endpoints.webhookTest');
-		this.endpointWebhookWaiting = config.getEnv('endpoints.webhookWaiting');
 
 		this.uniqueInstanceId = generateHostInstanceId(instanceType);
 	}
@@ -164,31 +152,13 @@ export abstract class AbstractServer {
 
 		// Setup webhook handlers before bodyParser, to let the Webhook node handle binary data in requests
 		if (this.webhooksEnabled) {
-			// Register a handler for active webhooks
-			this.app.all(
-				`/${this.endpointWebhook}/:path(*)`,
-				webhookRequestHandler(Container.get(ActiveWorkflowRunner)),
-			);
-
-			// Register a handler for waiting webhooks
-			this.app.all(
-				`/${this.endpointWebhookWaiting}/:path/:suffix?`,
-				webhookRequestHandler(Container.get(WaitingWebhooks)),
-			);
+			Container.get(ActiveWebhooks).registerHandler(this.app);
+			Container.get(WaitingWebhooks).registerHandler(this.app);
 		}
 
 		if (this.testWebhooksEnabled) {
-			const testWebhooks = Container.get(TestWebhooks);
-
 			// Register a handler for test webhooks
-			this.app.all(`/${this.endpointWebhookTest}/:path(*)`, webhookRequestHandler(testWebhooks));
-
-			// Removes a test webhook
-			// TODO UM: check if this needs validation with user management.
-			this.app.delete(
-				`/${this.restEndpoint}/test-webhook/:id`,
-				send(async (req) => testWebhooks.cancelTestWebhook(req.params.id)),
-			);
+			Container.get(TestWebhooks).registerHandler(this.app);
 		}
 
 		// Block bots from scanning the application
