+++ packages/cli/src/WorkflowHelpers.ts
@@ -29,13 +29,11 @@ import type {
 	IWorkflowExecutionDataProcess,
 } from '@/Interfaces';
 import { NodeTypes } from '@/NodeTypes';
-// eslint-disable-next-line import/no-cycle
 import { WorkflowRunner } from '@/WorkflowRunner';
 import config from '@/config';
 import type { WorkflowEntity } from '@db/entities/WorkflowEntity';
 import type { User } from '@db/entities/User';
 import omit from 'lodash/omit';
-// eslint-disable-next-line import/no-cycle
 import { PermissionChecker } from './UserManagement/PermissionChecker';
 import { UserService } from './services/user.service';
 import type { SharedWorkflow } from '@db/entities/SharedWorkflow';
