+++ packages/cli/src/WorkflowRunnerProcess.ts
@@ -218,7 +218,6 @@ class WorkflowRunnerProcess {
 				workflowData,
 				additionalData.userId,
 				options?.inputData,
-				options?.parentWorkflowId,
 			);
 			await sendToParentProcess('startExecution', { runData });
 			const executionId: string = await new Promise((resolve) => {
