+++ packages/cli/src/workflows/workflows.services.ts
@@ -25,8 +25,7 @@ import { TagService } from '@/services/tag.service';
 import type { IWorkflowDb, IWorkflowExecutionDataProcess } from '@/Interfaces';
 import { NodeTypes } from '@/NodeTypes';
 import { WorkflowRunner } from '@/WorkflowRunner';
-import * as WorkflowExecuteAdditionalData from '@/WorkflowExecuteAdditionalData';
-import { TestWebhooks } from '@/TestWebhooks';
+import { TestWebhooks } from '@/webhooks';
 import { whereClause } from '@/UserManagement/UserManagementHelper';
 import { InternalHooks } from '@/InternalHooks';
 import { WorkflowRepository } from '@/databases/repositories';
@@ -384,26 +383,12 @@ export class WorkflowsService {
 				startNodes.length === 0 ||
 				destinationNode === undefined)
 		) {
-			const workflow = new Workflow({
-				id: workflowData.id?.toString(),
-				name: workflowData.name,
-				nodes: workflowData.nodes,
-				connections: workflowData.connections,
-				active: false,
-				nodeTypes: Container.get(NodeTypes),
-				staticData: undefined,
-				settings: workflowData.settings,
-			});
-
-			const additionalData = await WorkflowExecuteAdditionalData.getBase(user.id);
-
 			const needsWebhook = await Container.get(TestWebhooks).needsWebhookData(
+				user.id,
 				workflowData,
-				workflow,
-				additionalData,
 				EXECUTION_MODE,
 				ACTIVATION_MODE,
-				sessionId,
+				sessionId!,
 				destinationNode,
 			);
 			if (needsWebhook) {
